+++ command/system/verify-install.go
@@ -16,10 +16,10 @@ import (
 var requiredFileRelative = "lc.yml"
 var requiredFile = fmt.Sprintf("/opt/project/%v", requiredFileRelative)
 
-// CmdVerifyLds will ensure lds disk mounting is functioning
-// This is mainly to address SAI-32.
-func CmdVerifyLds(c *cli.Context) error {
-	logrus.Debug("attempting to verify that the all lds components are functioning")
+// CmdVerifyInstall will ensure docker disk mounting is functioning
+// This is mainly to support troubleshooting of mac os x file sharing with the VM
+func CmdVerifyInstall(c *cli.Context) error {
+	logrus.Debug("attempting to verify that the all lc/docker components are functioning")
 	cwd, err := os.Getwd()
 	if err != nil {
 		return fmt.Errorf("could not find current working directory to verify repo: %q", err)
@@ -27,7 +27,7 @@ func CmdVerifyLds(c *cli.Context) error {
 
 	// first verify file exists locally
 	if _, err := os.Stat(filepath.Join(cwd, requiredFileRelative)); os.IsNotExist(err) {
-		logrus.Warnf("could not find '%v' in the current directory, skipping lds verification", requiredFileRelative)
+		logrus.Warnf("could not find '%v' in the current directory, skipping install verification", requiredFileRelative)
 		return nil
 	}
 
@@ -36,7 +36,7 @@ func CmdVerifyLds(c *cli.Context) error {
 	args := []string{"run", "--rm", "-v", volume, "--entrypoint=/bin/sh", "busybox", "-c", fileCheck}
 	cmd := exec.Command("docker", args...)
 	if err := helpers.RunCommand(cmd); err != nil {
-		return errors.New("It appears that your local disk is not mounted into the LDS VM. Try running 'lds reload' to fix, if that doesn't work see https://jira.lancope.local/browse/SAI-32 for more options on how to fix.")
+		return errors.New("It appears that your local disk is not mounted into the Docker-Daemon VM.")
 	}
 	return nil
 }